+++ create-vite-app/index.js
@@ -5,10 +5,11 @@ const { promisify } = require('util')
 const writeFile = promisify(fs.writeFile)
 const copyFile = promisify(fs.copyFile)
 const mkdir = promisify(fs.mkdir)
+const readdir = promisify(fs.readdir)
+const argv = require('minimist')(process.argv.slice(2))
 
 async function init() {
-  const targetDir = process.argv[2]
-
+  const targetDir = argv._[0]
   if (!targetDir) {
     console.error(
       `Error: Please specify target directory. Example: create-vite-app my-project`
@@ -17,15 +18,6 @@ async function init() {
   }
   const root = path.join(process.cwd(), targetDir || '')
   console.log(`Scaffolding project in ${root}...`)
-
-  const write = async (file, content) => {
-    const writePath = path.join(root, file)
-    if (content) {
-      await writeFile(writePath, content)
-    } else {
-      await copyFile(path.join(__dirname, `template/_${file}`), writePath)
-    }
-  }
 
   try {
     await mkdir(root)
@@ -38,18 +30,30 @@ async function init() {
     return
   }
 
-  await write('index.html')
-  await write('App.vue')
+  const templateDir = path.join(__dirname, `template-${argv.template || 'vue'}`)
+  const write = async (file, content) => {
+    const targetPath = path.join(root, file)
+    if (content) {
+      await writeFile(targetPath, content)
+    } else {
+      await copyFile(path.join(templateDir, file), targetPath)
+    }
+  }
 
-  const pkg = require('./template/_package.json')
-  pkg.name = path.basename(root)
+  const files = await readdir(templateDir)
+  for (const file of files.filter((f) => f !== 'package.json')) {
+    await write(file)
+  }
+
+  const pkg = require(path.join(templateDir, `package.json`))
+  pkg.name = targetDir
   await write('package.json', JSON.stringify(pkg, null, 2))
 
   console.log(`Done. Now run: `)
   console.log()
   console.log(`  cd ${path.relative(process.cwd(), root)}`)
   console.log(`  npm install (or \`yarn\`)`)
-  console.log(`  npx vite (or \`yarn vite\`)`)
+  console.log(`  npm run dev (or \`yarn dev\`)`)
   console.log()
 }