+++ src/node/utils/pathUtils.ts
@@ -1,6 +1,7 @@
 import { Context } from 'koa'
 import path from 'path'
 import slash from 'slash'
+import { InternalResolver } from '../resolver'
 
 export const queryRE = /\?.*$/
 export const hashRE = /\#.*$/
@@ -32,13 +33,51 @@ export const isStaticAsset = (file: string) => {
   return imageRE.test(file) || mediaRE.test(file) || fontsRE.test(file)
 }
 
+const timeStampRE = /(&|\?)t=\d+/
+const jsSrcFileRE = /\.(vue|jsx?|tsx?)$/
+
 /**
  * Check if a request is an import from js (instead of fetch() or ajax requests)
- * A request qualifies as long as it's not from page (no ext or .html).
- * this is because non-js files can be transformed into js and import json
- * as well.
+ * A request qualifies as long as it's from one of the supported js source file
+ * formats (vue,js,ts,jsx,tsx)
  */
-export const isImportRequest = (ctx: Context) => {
-  const referer = cleanUrl(ctx.get('referer'))
-  return /\.\w+$/.test(referer) && !referer.endsWith('.html')
+export const isImportRequest = (ctx: Context): boolean => {
+  if (!ctx.accepts('js')) {
+    return false
+  }
+  // strip HMR timestamps
+  const referer = ctx.get('referer').replace(timeStampRE, '')
+  return jsSrcFileRE.test(referer)
+}
+
+const moduleRE = /^[^\/\.]/
+const fileExtensionRE = /\.\w+$/
+
+export const resolveImport = (
+  importer: string,
+  id: string,
+  resolver: InternalResolver,
+  timestamp?: string
+): string => {
+  if (moduleRE.test(id)) {
+    return resolver.idToRequest(id) || `/@modules/${id}`
+  } else {
+    let { pathname, query } = resolveRelativeRequest(importer, id)
+    // append an extension to extension-less imports
+    if (!fileExtensionRE.test(pathname)) {
+      const file = resolver.requestToFile(pathname)
+      const indexMatch = file.match(/\/index\.\w+$/)
+      if (indexMatch) {
+        pathname = pathname.replace(/\/(index)?$/, '') + indexMatch[0]
+      } else {
+        pathname += path.extname(file)
+      }
+    }
+    // force re-fetch all imports by appending timestamp
+    // if this is a hmr refresh request
+    if (timestamp) {
+      query += `${query ? `&` : `?`}t=${timestamp}`
+    }
+    return pathname + query
+  }
 }